requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.124.2
h11==0.16.0
idna==3.11
psycopg[binary]==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
supabase>=2.10.0
Pillow==10.2.0
python-multipart==0.0.9
